requirements.txt
flask
flask-jwt-extended
flask-cors
flask-sqlalchemy
Pillow
psycopg2-binary
python-dotenv
python-barcode[images]